+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Hint at needing `noRemove: 1` as param when loop adding files (default
   behavior is to remove added files) in the `Files.tmp_upload` documentation.
 - `Connection` properly passes query parameters for post-ish requests now.
+- `Repository.add_file` no longer mangles it's query parameters as they are
+  inconsistently cased on the API level.
 
 ## [0.9.0]
 ### Added